Add tests and move things around
